+++ channel/serializers.py
@@ -1,0 +1,14 @@
+from rest_framework import serializers
+
+from .models import *
+
+
+class ChannelCreateSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Channel
+        fields = ['id']
+
+class ChannelSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Channel
+        fields = ['id', 'message', 'result']
