+++ channel/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+from .views import *
+
+urlpatterns = [
+    path('', ChannelCreateView.as_view(), name='channel_create'),
+    path('messages/<int:channel_id>/', SendMessageView.as_view(), name='send_message'),
+    path('results/', ChannelResultsView.as_view(), name='channel_results'),
+]
