+++ django_backend/urls.py
@@ -20,6 +20,8 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/v1/',include('testapp.urls')), # include 안에 자기 app 이름 적기
     path('api/v1/users/',include('user.urls')),
+    path('api/v1/channels/',include('channel.urls')),
+    path('api/v1/posts/',include('post.urls')),
     re_path(r'^swagger(?P<format>\\.json|\\.yaml)$', schema_view.without_ui(cache_timeout=0), name='schema-json'),
     re_path(r'^swagger/$', schema_view.with_ui('swagger', cache_timeout=0), name='schema-swagger-ui'),
     re_path(r'^redoc/$', schema_view.with_ui('redoc', cache_timeout=0), name='schema-redoc'),
